+++ projects/ngx-uploader-directive/src/public-api.ts
@@ -2,6 +2,7 @@
  * Public API Surface of lib-file-uploader
  */
 
+export * from './lib/configs/config';
 export * from './lib/models/ngx-uploader-directive-models';
 export * from './lib/directives/ngx-uploader-drop.directive';
 export * from './lib/directives/ngx-uploader-select.directive';