+++ src/app/app.component.ts
@@ -10,25 +10,30 @@ import { IUploadOptions, ISelectedFile, IUploadInput, IUploadOutput } from 'ngx-
 
 export class AppComponent {
   title = 'ngx-uploader-directive';
+  fileId: string;
   options: IUploadOptions;
   formData: FormData;
   files: Array<ISelectedFile>;
   uploadInput: EventEmitter<IUploadInput>;
   dragOver: boolean;
-  uploadUrl = 'http://192.168.0.224:8099/api/blocklists/uploadblockednumberfile';
+  uploadUrl = 'Api Url';
 
   /**
    * Default Constructor
    */
   constructor() {
-    this.options = { requestConcurrency: 1, maxFilesToAddInSingleRequest: 10, maxFileUploads: 5, maxFileSize: 1000000, logs: true };
+    this.options = { requestConcurrency: 1, maxFilesToAddInSingleRequest: 2, maxFileUploads: 10, maxFileSize: 1000000, logs: true };
     this.files = new Array<ISelectedFile>();
     this.uploadInput = new EventEmitter<IUploadInput>();
     this.formData = new FormData();
   }
 
+  /**
+   * Upload output events.
+   * @param output IUploadOutput Model on output.
+   */
   onUploadOutput(output: IUploadOutput): void {
-    // console.log(output);
+    console.log(output);
     switch (output.type) {
       case 'init':
         this.files = new Array<ISelectedFile>();
@@ -38,21 +43,19 @@ export class AppComponent {
         // this.startUpload();
         break;
       case 'addedToQueue':
-        if (typeof output.file !== 'undefined') {
-          this.files.push(output.file);
-          // console.log(this.files);
-        }
+        this.files = this.files.concat(output.files);
+        console.log(this.files);
+        break;
+      case 'start':
+        // uploading start  
         break;
       case 'uploading':
-        if (typeof output.file !== 'undefined') {
-          // update current data in files array for uploading file
-          const index = this.files.findIndex((file) => typeof output.file !== 'undefined' && file.id === output.file.id);
-          this.files[index] = output.file;
-        }
+        this.files = this.updateFiles(this.files, output.files, 'UPDATE');
+        console.log(this.files);
         break;
       case 'removed':
-        // remove file from array when removed
-        this.files = this.files.filter((file: ISelectedFile) => file !== output.file);
+        this.files = this.updateFiles(this.files, output.files, 'REMOVE');
+        console.log(this.files);
         break;
       case 'dragOver':
         this.dragOver = true;
@@ -62,17 +65,48 @@ export class AppComponent {
         this.dragOver = false;
         break;
       case 'done':
-        // The file is downloaded
+        // The files are uploaded
+        this.files = this.updateFiles(this.files, output.files, 'UPDATE');
+        console.log(this.files);
         break;
     }
   }
 
+  /**
+   * Update files on output events
+   * @param currentFiles Current Files Array
+   * @param updatedFiles Updated Files Array
+   * @param action Remove or Update
+   */
+  updateFiles(currentFiles: Array<ISelectedFile>, updatedFiles: Array<ISelectedFile>, action: 'REMOVE' | 'UPDATE') {
+    if (updatedFiles !== undefined) {
+      if (action === 'UPDATE') {
+        updatedFiles.forEach(updateFile => {
+          currentFiles.forEach(
+            (currentFile, currentFileIndex, currentFilesArray) => {
+              if (currentFile.name === updateFile.name) {
+                currentFilesArray[currentFileIndex] = updateFile;
+              }
+            }
+          );
+        });
+      } else if (action === 'REMOVE') {
+        currentFiles = currentFiles.filter((file) => file.requestId !== updatedFiles[0].requestId);
+      }
+    }
+    return currentFiles;
+  }
+
+  /**
+   * Start Upload
+   */
   startUpload(): void {
     this.formData.append('fileHasHeader', 'false');
     this.formData.append('delimiter', ',');
 
     const event: IUploadInput = {
       type: 'uploadAll',
+      inputReferenceNumber: Math.random(),
       url: this.uploadUrl,
       method: 'POST',
       formData: this.formData,
@@ -82,15 +116,27 @@ export class AppComponent {
     this.uploadInput.emit(event);
   }
 
-  cancelUpload(id: string): void {
-    this.uploadInput.emit({ type: 'cancel', id });
+  /**
+   * Cancel file uploads.
+   * @param requestId RequestId.
+   */
+  cancelUpload(requestId: string): void {
+    this.uploadInput.emit({ type: 'cancel', inputReferenceNumber: Math.random(), id: requestId });
   }
 
-  removeFile(id: string): void {
-    this.uploadInput.emit({ type: 'remove', id });
+  /**
+   * Remoce files.
+   * @param requestId Request id
+   */
+  removeFile(requestId: string): void {
+    console.log(requestId);
+    this.uploadInput.emit({ type: 'remove', inputReferenceNumber: Math.random(), id: requestId });
   }
 
+  /**
+   * Remoce all file uploads.
+   */
   removeAllFiles(): void {
-    this.uploadInput.emit({ type: 'removeAll' });
+    this.uploadInput.emit({ type: 'removeAll', inputReferenceNumber: Math.random(), });
   }
 }