+++ CHANGELOG.md
@@ -4,13 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.2] - 2020-05-03
+
+### Fixed
+- fixed multiple events response
+- Fixed on upload first event was error
+
 ## [1.1.0] - 2020-04-29
 
 ### Added
 
 - Facility to send headers with request
 
-### Changed
+### Fixed
 - Fixed on removing file it will send file and the id of deleted file
 - Fixed on cancel uploading file it will send file and the id of cencelled file.